Refactor PlacementType to lib-common
