+++ frontend/src/citizen-frontend/applications/editor/service-need/ServiceNeedSection.tsx
@@ -2,11 +2,11 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-import { ServiceNeedOptionPublicInfo } from 'citizen-frontend/applications/api'
 import {
   ApplicationStatus,
   ApplicationType
 } from 'lib-common/api-types/application/enums'
+import { ServiceNeedOptionPublicInfo } from 'lib-common/api-types/serviceNeed/common'
 import { UpdateStateFn } from 'lib-common/form-state'
 import LocalDate from 'lib-common/local-date'
 import HorizontalLine from 'lib-components/atoms/HorizontalLine'
