+++ frontend/src/employee-frontend/api/child/placements.ts
@@ -8,7 +8,7 @@ import { Placement } from '../../types/child'
 import { UUID } from '../../types'
 import { JsonOf } from 'lib-common/json'
 import LocalDate from 'lib-common/local-date'
-import { PlacementType } from 'lib-customizations/types'
+import { PlacementType } from 'lib-common/api-types/serviceNeed/common'
 
 export interface PlacementCreate {
   childId: UUID
