+++ frontend/src/employee-frontend/api/child/service-needs.ts
@@ -6,9 +6,12 @@ import { Failure, Result, Success } from 'lib-common/api'
 import { client } from '../client'
 import { ServiceNeedOption } from '../../types/child'
 import { JsonOf } from 'lib-common/json'
-import LocalDate from 'lib-common/local-date'
-import { PlacementType } from 'lib-customizations/types'
-import { UUID } from 'lib-common/types'
+import LocalDate from '../../../lib-common/local-date'
+import { UUID } from '../../types'
+import {
+  PlacementType,
+  ServiceNeedOptionPublicInfo
+} from '../../../lib-common/api-types/serviceNeed/common'
 
 export interface ServiceNeedCreateRequest {
   placementId: UUID
@@ -58,12 +61,6 @@ export async function getServiceNeedOptions(): Promise<
     .get<JsonOf<ServiceNeedOption[]>>('/service-needs/options')
     .then((res) => Success.of(res.data))
     .catch((e) => Failure.fromError(e))
-}
-
-export interface ServiceNeedOptionPublicInfo {
-  id: UUID
-  name: string
-  validPlacementType: PlacementType
 }
 
 export async function getServiceNeedOptionPublicInfos(
