+++ frontend/src/employee-frontend/components/child-information/placements/CreatePlacementModal.tsx
@@ -17,11 +17,11 @@ import { createPlacement } from '../../../api/child/placements'
 import Select from '../../../components/common/Select'
 import { FixedSpaceColumn } from 'lib-components/layout/flex-helpers'
 import { AlertBox } from 'lib-components/molecules/MessageBoxes'
+import { PlacementType } from 'lib-common/api-types/serviceNeed/common'
 import { Unit } from '../../../types/unit'
 import { getDaycares } from '../../../api/unit'
 import { useRestApi } from '../../../../lib-common/utils/useRestApi'
 import DateRange from 'lib-common/date-range'
-import { PlacementType } from 'lib-customizations/types'
 import { placementTypes } from 'lib-customizations/employee'
 import colors from 'lib-customizations/common'
