+++ frontend/src/employee-frontend/types/invoicing.ts
@@ -2,10 +2,10 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
+import { PlacementType } from 'lib-common/api-types/serviceNeed/common'
 import DateRange from 'lib-common/date-range'
 import { JsonOf } from 'lib-common/json'
 import LocalDate from 'lib-common/local-date'
-import { PlacementType } from 'lib-customizations/types'
 import { UUID } from '../types'
 import { FeeAlterationType } from '../types/fee-alteration'
 import { Income, IncomeEffect } from '../types/income'
