+++ frontend/src/employee-frontend/utils/placements.ts
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-import { PlacementType } from 'lib-customizations/types'
+import { PlacementType } from 'lib-common/api-types/serviceNeed/common'
 
 const partDayPlacementTypes: readonly PlacementType[] = [
   'DAYCARE_PART_TIME',
