+++ frontend/src/employee-mobile-frontend/api/attendances.ts
@@ -14,7 +14,7 @@ import {
   deserializeAbsence
 } from 'lib-common/api-types/child/Absences'
 import FiniteDateRange from 'lib-common/finite-date-range'
-import { PlacementType } from 'lib-customizations/types'
+import { PlacementType } from 'lib-common/api-types/serviceNeed/common'
 
 export interface DepartureInfoResponse {
   absentFrom: CareType[]
