+++ frontend/src/lib-common/api-types/serviceNeed/common.ts
@@ -4,7 +4,27 @@
 
 import { UUID } from 'lib-common/types'
 
+export interface ServiceNeedOptionPublicInfo {
+  id: UUID
+  name: string
+  validPlacementType: PlacementType
+}
+
 export interface ServiceNeedOptionSummary {
   id: UUID
   name: string
 }
+
+export type PlacementType =
+  | 'CLUB'
+  | 'DAYCARE'
+  | 'DAYCARE_PART_TIME'
+  | 'DAYCARE_FIVE_YEAR_OLDS'
+  | 'DAYCARE_PART_TIME_FIVE_YEAR_OLDS'
+  | 'PRESCHOOL'
+  | 'PRESCHOOL_DAYCARE'
+  | 'PREPARATORY'
+  | 'PREPARATORY_DAYCARE'
+  | 'TEMPORARY_DAYCARE'
+  | 'TEMPORARY_DAYCARE_PART_DAY'
+  | 'SCHOOL_SHIFT_CARE'
